MusicTypewriter goes on
MusicTypewriter assimilate Punched Tapes for Flexowriter. Two alternatives are available known as MutualKeying. Short and less complex is the Consonantal Alphabet on a limited Octave range. More complex, but better is the limited Octave range by using VisualBraille as a font. The difference: Each mapped or assigned letter is punched multiple times and can be seen as a chord.